added the file-service module; added portal module; refactored post and thumbs; refactored tests

+++ fludland-service/src/main/java/org/fludland/services/ThumbService.java
@@ -1,6 +1,6 @@
 package org.fludland.services;
 
 public interface ThumbService {
-    void putPostThumb(long postId, long userId);
+    void putPostThumb(long postId, int userId);
     long getTotalPostThumbs(long postId);
 }
